fix to not explicitely link against ncurses
There is no need to explicitely link against libncurses because linking against
libreadline will pull in the libraries it depends on.
One should also note, that ncurses libraries can be build seperately, which
means that readline (if configured to use curses instead of termcap) will be
only linked against libtinfo.